Fixed the wall stuck bug
Fixed the bug where the player character stuck to the wall. It was an
issue with the Ground script and incorrect ground detection on a wall.
And also an issue with the frictionless material as well.
Where the friction it was applying allowed the rigid body to stick to
wall.