Fixed Game breaking bug with key
The key was not interacting with Door3. Because the distance it used to
check if it was near the target checked to see if the dis < 2. During
the game even if we placed the key as near the door as possible it only
had a dis = 3.92 . Thus never really fulfilling that condition. So I
fixed it by changing the codition check to dis < 4. Thus allowing the
game to proceed ahead.

Apparently the mirror and reaching the shroom below still do not end the
game. Need to investigate.